S12P11A204/backend/requirements.txt
alembic==1.14.0
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
blinker==1.9.0
boto3==1.36.2
botocore==1.36.4
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
config==0.5.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==4.2.18
django-background-tasks==1.2.8
django-cors-headers==4.3.1
django-environ==0.12.0
django-extensions==3.2.3
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
environ==1.0
gevent==24.11.1
greenlet==3.1.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
kombu==5.4.2
led==0.0.2
Mako==1.3.8
MarkupSafe==3.0.2
numpy==2.0.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
pillow==11.1.0
prompt_toolkit==3.0.48
psycopg==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.31.0
rest-framework-simplejwt==0.0.2
s3transfer==0.11.1
scikit-learn==1.6.1
scipy==1.13.1
service-identity==24.2.0
six==1.17.0
smbus2==0.5.0
SQLAlchemy==2.0.37
sqlparse==0.5.3
threadpoolctl==3.5.0
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
websocket==0.2.1
websockets==14.2
Werkzeug==3.1.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
